[homework-1] finish qr + plot fcts, start report
Finish coding both QRfactorise and QRsolve,
and add plot_complexity to show that the algorithm is
in O(n^3).
Still need to add some robustness to the code.

Also start the report: add solver algorithm,
complexity for QRsolve and total,
still need to add a more detailed version for QRfactorise.
Also need to add plots.